requirements-vendor.txt
Django==1.9.2
python-social-auth==0.2.14
requests==2.3
facebook-sdk==0.4.0
